+++ changelog.md
@@ -4,6 +4,10 @@
 
 * Add new constants for emitter
 
+### Bug Fixes
+
+* Fix bug where ganglion sample channel data when scale was false produced array of 8 values instead of 4
+
 # v0.2.3
 
 ### New Features